Move stats.rs to the component module
