+++ src/component/stats.rs
